Annoying game
Keeps asking the user if we are there yet till they say yes

checks for yes in a string at all, for ex. the user could say "Yes we are there" but using a normal (answer == "yes") wont recognize the yes. So by adding the indexof method, it will check for the word "yes" at any point in a string.